feat(queries): add multi-line queries support
Features some refactoring to support this feature, to remove the
dependency of compiling a complex reg-exp I preferred keeping the
output of the findQueries function untransformed, so we can use that
string to find and replace the query with the serialized query.

It adds a boolean to check if we are capturing a multi-line query or not, and
some logic to complete the feature.

I've tested both multi-line and single-line queries to ensure that I did
not break basic functionality.