fix(misc): remove newline insertion at each serialization
Everytime I would update a document that has the Dataview serialization
queries, it would insert an additional newline at the end of the
serialization.

If you do the modification several times, the document would create a
lot of useless newlines, so I just suggest to remove it and keep the
document clean.